# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API",
        "title": "Streaming API",
        "contact": {
            "name": "KauêTTS",
            "url": "https://github.com/KaueTTS"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/v1/contents": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lista filmes ou séries usando a integração com a TMDB API",
                "tags": [
                    "contents"
                ],
                "summary": "Listar conteúdos",
                "parameters": [
                    {
                        "enum": [
                            "movie",
                            "tv"
                        ],
                        "type": "string",
                        "description": "Tipo do conteúdo",
                        "name": "type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Número da página",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Ordenação usada pela TMDB. Exemplo: popularity.desc",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "IDs de gêneros separados por vírgula",
                        "name": "with_genres",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Idioma da resposta. Exemplo: pt-BR",
                        "name": "language",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Ano de lançamento",
                        "name": "year",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto_content.ContentResponseDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto_shared.ErrorDto"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto_shared.ErrorDto"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/dto_shared.ErrorDto"
                        }
                    }
                }
            }
        },
        "/v1/contents/search": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Busca filmes ou séries usando a integração com a TMDB API",
                "tags": [
                    "contents"
                ],
                "summary": "Buscar conteúdos",
                "parameters": [
                    {
                        "enum": [
                            "movie",
                            "tv"
                        ],
                        "type": "string",
                        "description": "Tipo do conteúdo",
                        "name": "type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Termo pesquisado",
                        "name": "query",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Número da página",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Idioma da resposta. Exemplo: pt-BR",
                        "name": "language",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto_content.ContentResponseDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto_shared.ErrorDto"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto_shared.ErrorDto"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/dto_shared.ErrorDto"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto_content.ContentDto": {
            "type": "object",
            "properties": {
                "backdrop_path": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "external_id": {
                    "type": "integer"
                },
                "genre_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "original_language": {
                    "type": "string"
                },
                "original_title": {
                    "type": "string"
                },
                "popularity": {
                    "type": "number"
                },
                "poster_path": {
                    "type": "string"
                },
                "release_date": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "vote_average": {
                    "type": "number"
                },
                "vote_count": {
                    "type": "integer"
                }
            }
        },
        "dto_content.ContentResponseDto": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto_content.ContentDto"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/dto_shared.PaginationDto"
                }
            }
        },
...
        "dto_shared.DetailErrorDto": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "dto_shared.ErrorDto": {
            "type": "object",
            "properties": {
                "codeMessage": {
                    "type": "string"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto_shared.DetailErrorDto"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "dto_shared.PaginationDto": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "pageCount": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}